Extract editor extensions into editor-extensions.ts (editor.ts 221 -> 174 lines)
Moved the TipTap extension configuration array into a dedicated
buildEditorExtensions() function, keeping editor.ts focused on
initialization and wiring.
